Initial version for Gerrit 2.14
This plugin allows to enforce rate limits in Gerrit. Rate limits define
the maximum request rate for users in a given group.

This plugin is a fork of rate limits feature of the quota plugin
introduce in [1]. The main reason for forking/extracting the rate limits
out of the quota was to be able to use rate limits without the overhead
of quota which is calculating size on disk even if no quota are set.

[1]https://gerrit-review.googlesource.com/c/plugins/quota/+/106976

Change-Id: I00f4c3af55ae540d644e2ca021b5d9758c98b829