Introduce a configuration to define time lapse for rate limiter
Before this change, rate limiter was executed every hour. Sometimes,a
sudden high memory demand on a short period of time happens caused by
high number of clones and fetches. Also, the user get blocked from
fetching information due to the limit per hour.

This change introduces an optional configurable field "timelapse". This
field defines a duration in minuite supported by the rate-limiter
plugin. This helps address the potential problem of high memory demand.
The value of this field is limited to 59 or less than an hour. If this
field is not defined, a default value of 1 hour is established.

Change-Id: I5cdbb931d676c7630178d0b8ba0245e3c8b774a8