Introduce a configuration to define time lapse for rate limiter
Before this change, rate limiter was executed every hour. Sometimes,a
sudden high memory demand on a short period of time happens caused by
high number of clones and fetches. Also, the user get blocked from
fetching information due to the limit per hour.

This change introduces an optional configurable field "timelapse". This
field defines a window supported by the rate-limiter plugin. This new
field defines the time for rate-limiter in minutes. If this field is not
defined, a default value of 1 hour is established.

Change-Id: I5cdbb931d676c7630178d0b8ba0245e3c8b774a8